rename spec file
